+++ Snake Game/Point.h
@@ -6,15 +6,15 @@
 class Point
 {
 public:
-	int x;
-	int y;
+	float x;
+	float y;
 
 	Point() {
 		x = 0;
 		y = 0;
 	}
 
-	Point(int x, int y) {
+	Point(float x, float y) {
 		this->x = x;
 		this->y = y;
 	}
